server/authentication/tests/test_views.py
from authentication.views import CsrfTokenViewSet
from django.test import RequestFactory
from rest_framework.permissions import AllowAny
from rest_framework.response import Response

"""
Code Analysis

Main functionalities:
The CsrfTokenViewSet class is a viewset that retrieves the CSRF token and sets it as a cookie. It is used to ensure that requests made to the server are legitimate and not forged. This class is designed to be used with Django and the Django Rest Framework.

Methods:
- get_csrf_token: This method retrieves the CSRF token and sets it as a cookie. It is decorated with the ensure_csrf_cookie decorator to ensure that the CSRF cookie is set correctly.

Fields:
- permission_classes: This field is set to AllowAny, which means that any user can access the CSRF token. This is necessary because the CSRF token is required for all requests to the server, including those made by unauthenticated users.
"""


class TestCsrfTokenViewSet:
    # Tests that the viewset behaves correctly when there are no edge cases.
    def test_edge_case_none(self):
        """
        Tests that the viewset behaves correctly when there are no edge cases.
        """
        viewset = CsrfTokenViewSet()
        request = RequestFactory().get("/")
        response = viewset.get_csrf_token(request)
        assert response.status_code == 200

    # Tests that the viewset has the correct permission classes.
    def test_general_behavior_permission_classes(self):
        """
        Tests that the viewset has the correct permission classes.
        """
        viewset = CsrfTokenViewSet()
        assert viewset.permission_classes == [AllowAny]

    # Tests that the viewset has the correct action.
    def test_general_behavior_action(self):
        """
        Tests that the viewset has the correct action.
        """
        viewset = CsrfTokenViewSet()
        assert viewset.get_csrf_token.mapping == {"get": "get_csrf_token"}

    # Tests that the viewset returns a response.
    def test_general_behavior_response(self):
        """
        Tests that the viewset returns a response.
        """
        viewset = CsrfTokenViewSet()
        request = RequestFactory().get("/")
        response = viewset.get_csrf_token(request)
        assert isinstance(response, Response)
